Add workflow to test code coverage and generate coverage report (#24)
* Add workflow to test code coverage and generate coverage report

This commit adds a GitHub Actions workflow (`coverage.yml`) . The workflow runs on pull requests targeting the `main` branch and consists of a job named `test-and-coverage`. The job runs tests with coverage enabled, generates an HTML coverage report, and uploads the report as an artifact. Additionally, the workflow extracts the pull request title and description and comments on the pull request with the coverage results.

* Add utransport test cases

* Increase code coverage

With these new tests, we were able to increase code coverage to 98%. Along with this, we provided some linting improvements, as well as fixed a few methods that were implemented incorrectly in a previous PR.

* Update pyproject.toml

Current system does not support tool.coverage.report, so I have removed it for now. Those exclusion cases will be covered in the coverage.rc file

* Add steps to comment on the PR with the coverage report

* Remove comment on PR step

---------

Co-authored-by: Matthew D'Alonzo <matthew.dalonzo@gm.com>